New parser module for code-share with make_layout / collection
